+++ source/extensions/octi.lab_tasks/octi/lab_tasks/tasks/locomotion/fetching/config/unitree_go1/flat_env_cfg.py
@@ -5,19 +5,19 @@
 
 from omni.isaac.lab.utils import configclass
 
-from .rough_env_cfg import UnitreeA1FetchingRoughCfg
+from .rough_env_cfg import UnitreeGo1RoughEnvCfg
 
 
 @configclass
-class UnitreeA1FetchingFlatCfg(UnitreeA1FetchingRoughCfg):
+class UnitreeGo1FlatEnvCfg(UnitreeGo1RoughEnvCfg):
     def __post_init__(self):
         # post init of parent
         super().__post_init__()
 
         # override rewards
-        # self.rewards.flat_orientation_l2.weight = -5.0
-        self.rewards.dof_torques_l2.weight = -2.5e-5
-        self.rewards.feet_air_time.weight = 5
+        self.rewards.flat_orientation_l2.weight = -2.5
+        self.rewards.feet_air_time.weight = 0.25
+
         # change terrain to flat
         self.scene.terrain.terrain_type = "plane"
         self.scene.terrain.terrain_generator = None
@@ -26,4 +26,3 @@ class UnitreeA1FetchingFlatCfg(UnitreeA1FetchingRoughCfg):
         self.observations.policy.height_scan = None
         # no terrain curriculum
         self.curriculum.terrain_levels = None
-